move scopes to its own file
