+++ K6/api/tests/authentication/consent/consent-commons.js
@@ -1,5 +1,6 @@
 import http from "k6/http";
 import { parseCsvData } from "../../../../helpers.js";
+import { ENDUSER_SCOPE } from "../../../../scopes.js";
 
 /**
  * Shared helpers for the consent request events tests and their test data
@@ -17,15 +18,6 @@ import { parseCsvData } from "../../../../helpers.js";
 
 const TESTDATA_BASE_URL =
     "https://raw.githubusercontent.com/Altinn/altinn-platform-validation-tests/refs/heads/main/K6/testdata/authentication/consent";
-
-export const ConsentScope = {
-    READ: "altinn:consentrequests.read",
-    WRITE: "altinn:consentrequests.write",
-    // Maskinporten looks up consent before issuing a token (org token).
-    LOOKUP: "altinn:maskinporten/consent.read",
-};
-
-export const ENDUSER_SCOPE = "altinn:portal/enduser";
 
 /**
  * The organizations that receive (and therefore hold) consents.
@@ -64,35 +56,29 @@ export function getLookupConsents(env) {
 }
 
 /**
- * Token options for a consentee organization (enterprise token).
- * orgNo is optional so the generator can be built once and have its org set
- * per iteration via setTokenGeneratorOptions.
+ * Token options for a consentee organization (enterprise token), for a given
+ * org and scope. Set per iteration via setTokenGeneratorOptions.
  * @param {string} env
- * @param {string} [orgNo]
- * @param {string} scopes One of {@link ConsentScope}.
+ * @param {string} orgNo
+ * @param {string} scopes
  * @returns {Map<string, string|number>}
  */
 export function getEnterpriseTokenOpts(env, orgNo, scopes) {
     const opts = getBaseTokenOpts(env, scopes);
-    if (orgNo !== undefined) {
-        opts.set("orgNo", orgNo);
-    }
+    opts.set("orgNo", orgNo);
     return opts;
 }
 
 /**
- * Token options for a consenter person (personal token).
- * partyUuid is optional so the generator can be built once and have its
- * identity set per iteration via setTokenGeneratorOptions.
+ * Token options for a consenter person (personal token), for a given person.
+ * Set per iteration via setTokenGeneratorOptions.
  * @param {string} env
- * @param {string} [partyUuid]
+ * @param {string} partyUuid
  * @returns {Map<string, string|number>}
  */
 export function getPersonalTokenOpts(env, partyUuid) {
     const opts = getBaseTokenOpts(env, ENDUSER_SCOPE);
-    if (partyUuid !== undefined) {
-        opts.set("partyuuid", partyUuid);
-    }
+    opts.set("partyuuid", partyUuid);
     return opts;
 }
